Refactor: Move ImGui-related files to `kfx/imgui` directory
